feat: ViewController 병합
